requirements.txt
#python 3.7

numpy
scipy
tensorflow    # version 2.3
matplotlib
tf-nightly
seaborn

randomgen

pyyaml

PyQt5 # for plotting with gui